backend/requirements.txt
fastapi
uvicorn
pillow
requests
python-multipart 
